refactor(lib-nexus): move files
